fix(web-app): do not display hidden library nav  items in side nav
